Created one tasks page file to include to all users tasks page and Tasks Controller to handle tasks page user interactions
