+++ step_system/app/Views/partials/planning/plan-nav.php
@@ -64,8 +64,8 @@ $is_procurement_page = in_array($current_path, $procurement_pages);
             </li>
 
             <li class="menu">
-                <a href="<?= base_url('planning/tasks') ?>"
-                   class="dropdown-toggle <?= $current_path === 'planning/tasks' ? 'active text-white fw-bold shadow-text' : '' ?>">
+                <a href="<?= base_url('/tasks') ?>"
+                   class="dropdown-toggle <?= $current_path === '/tasks' ? 'active text-white fw-bold shadow-text' : '' ?>">
                     <div>
                         <img src="<?= base_url('assets/images/icon-tasks.svg') ?>" width="24" height="24" alt="checklist">
                         <span class="ms-2 shadow-text">Tasks</span>
