+++ step_system/public/assets/js/tasks_page/tasks.js
@@ -34,7 +34,7 @@ document.addEventListener('DOMContentLoaded', function () {
         modalActionButtons.style.display = 'block';
         modalStatusDisplay.style.display = 'none';
 
-        fetch(`/planning/tasks/details/${taskId}`)
+        fetch(`/tasks/details/${taskId}`)
             .then(response => {
                 if (!response.ok) throw new Error('Network response was not ok');
                 return response.json();
@@ -101,7 +101,7 @@ document.addEventListener('DOMContentLoaded', function () {
                     }
                 });
 
-                fetch('/planning/tasks/update-status', {
+                fetch('/tasks/update-ppmp-status', {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json',
@@ -138,4 +138,19 @@ document.addEventListener('DOMContentLoaded', function () {
 
     approveBtn.addEventListener('click', handleStatusUpdate);
     rejectBtn.addEventListener('click', handleStatusUpdate);
+
+    c2 = $('#style-2').DataTable({
+        "dom": "<'dt--top-section'<'row'<'col-12 col-sm-6 d-flex justify-content-sm-start justify-content-center'f><'col-12 col-sm-6 d-flex justify-content-sm-end justify-content-center mt-sm-0 mt-3'>>>" +
+    "<'table-responsive'tr>" +
+    "<'dt--bottom-section d-sm-flex justify-content-sm-between text-center'<'dt--pages-count  mb-sm-0 mb-3'i><'dt--pagination'p>>",
+        "oLanguage": {
+            "oPaginate": { "sPrevious": '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-arrow-left"><line x1="19" y1="12" x2="5" y2="12"></line><polyline points="12 19 5 12 12 5"></polyline></svg>', "sNext": '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-arrow-right"><line x1="5" y1="12" x2="19" y2="12"></line><polyline points="12 5 19 12 12 19"></polyline></svg>' },
+            "sInfo": "Showing page _PAGE_ of _PAGES_",
+            "sSearch": '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-search"><circle cx="11" cy="11" r="8"></circle><line x1="21" y1="21" x2="16.65" y2="16.65"></line></svg>',
+            "sSearchPlaceholder": "Search...",
+            "sEmptyTable": "No tasks at the moment. Lucky you hehe :>"
+        },
+        "pageLength": 10,
+        "order": []
+    });
 });
